backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
websockets==12.0

# ── Database ──────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ── Auth ──────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.9

# ── Task Queue ────────────────────────────────────────────────
celery[redis]==5.3.6
redis==5.0.1
celery-redbeat==2.1.1

# ── Docker SDK ────────────────────────────────────────────────
docker==7.0.0

# ── PDF ───────────────────────────────────────────────────────
weasyprint==62.1
jinja2==3.1.3

# ── HTTP ──────────────────────────────────────────────────────
httpx==0.26.0

# ── Validation ────────────────────────────────────────────────
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
email-validator==2.1.0.post1

# ── Monitoring ────────────────────────────────────────────────
prometheus-fastapi-instrumentator==6.1.0
structlog==24.1.0

# ── Security ──────────────────────────────────────────────────
cryptography==42.0.2
argon2-cffi==23.1.0
reportlab>=4.0.0
sherlock-project==0.16.0
numpy==1.26.4
pandas>=2.0.0
openpyxl>=3.0.0
python-docx>=1.1.0
